requirements.txt
Babel==2.5.3
boto==2.48.0
boto3==1.5.26
botocore==1.8.40
certifi==2018.4.16
chardet==3.0.4
dj-database-url==0.4.2
Django==1.11.3
django-bootstrap-static==3.3.7.1
django-bootstrap3==9.1.0
django-datetime-widget==0.9.3
django-debug-toolbar==1.9.1
django-extensions==1.9.6
django-filter==1.1.0
django-geojson==2.11.0
django-leaflet==0.22.0
django-phonenumber-field==2.0.0
django-progressive-web-app==0.1.1
django-storages==1.6.5
docutils==0.14
gunicorn==19.7.1
idna==2.6
jmespath==0.9.3
jsonfield==2.0.2
numpy==1.14.0
olefile==0.44
pandas==0.22.0
phonenumberslite==8.8.11
Pillow==4.3.0
psycopg2==2.7.4
python-dateutil==2.6.1
pytz==2017.3
requests==2.18.4
s3transfer==0.1.12
six==1.11.0
sqlparse==0.2.4
urllib3==1.22
whitenoise==3.3.1
